refactor(tests): split fanout health fixtures
Move local Git, clock, readiness, and dispatch fixture helpers out of the health-event behavior suite so both files stay below the repository pure-LOC ceiling without changing coverage.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: preserve the production projector and dispatcher assertions exactly; this is a size-only test organization change.

Rejected: waive a two-line size overage | the no-excuse architecture gate applies to test modules too.

Confidence: high

Scope-risk: narrow

Directive: keep shared setup in the non-discovery support module and behavior assertions in the named test module.

Tested: 8 fanout health-event tests; Ruff; compileall; git diff --check; LSP; behavior file 220 pure LOC and helper 73 pure LOC.

Not-tested: full frozen suite follows on the new clean tree.
